% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adwords.metrics}
\alias{adwords.metrics}
\title{Adwords Report Metrics}
\format{A data frame with 7830 rows and 11 variables}
\usage{
adwords.metrics
}
\description{
A table containing all the Adwords report metrics for the 201609, 201607 and 201605 API versions.
}
\keyword{MC-Fold}
\keyword{RNA}
\keyword{datasets}

