% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adwords.reports}
\alias{adwords.reports}
\title{Adwords Reports}
\format{A data frame with 132 rows and 3 variables}
\usage{
adwords.reports
}
\description{
A table containing all the Adwords reports for the 201702, 201609 and 201607 API versions.
This data is extracted from the files in the RAdwords package.
}
\keyword{Adwords}
\keyword{Report}
\keyword{datasets}
