% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.xml.R
\name{as.xml}
\alias{as.xml}
\alias{as.xml.NULL}
\alias{as.xml.list}
\alias{as.xml.logical}
\alias{as.xml.predicate}
\alias{as.xml.search.parameter}
\alias{as.xml.language}
\alias{as.xml.location}
\alias{as.xml.network}
\alias{as.xml.related.to.query}
\alias{as.xml.related.to.url}
\alias{as.xml.keyword.estimate.request}
\alias{as.xml.adgroup.estimate.request}
\alias{as.xml.campaign.estimate.request}
\alias{as.xml.max.cpc}
\alias{as.xml.criterion}
\title{XML Encoding for Adwords Service Requests}
\usage{
as.xml(x)

\method{as.xml}{NULL}(x)

\method{as.xml}{list}(x)

\method{as.xml}{logical}(x)

\method{as.xml}{predicate}(x)

\method{as.xml}{search.parameter}(x)

\method{as.xml}{language}(x)

\method{as.xml}{location}(x)

\method{as.xml}{network}(x)

\method{as.xml}{related.to.query}(x)

\method{as.xml}{related.to.url}(x)

\method{as.xml}{keyword.estimate.request}(x)

\method{as.xml}{adgroup.estimate.request}(x)

\method{as.xml}{campaign.estimate.request}(x)

\method{as.xml}{max.cpc}(amount)

\method{as.xml}{criterion}(criterion)
}
\arguments{
\item{x}{Data to convert to XML.}
}
\value{
An XMLNode encoding the data
}
\description{
Encodes the search parameters of an Adwords service request into XMl.
Allow the conversion of the data to a SOAP message for the Adwords service.
}
\examples{
TODO
}
