% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize a string}
\usage{
capitalize(x)
}
\arguments{
\item{x}{string to capitalize}
}
\value{
Capitalized string
}
\description{
Puts the first letter of a string in uppercase.
Used in some of the parse functions.
Inspiration drawn from the examples of the \link{toupper} function.
}
\examples{
capitalize("capital")
}
\seealso{
\link{parse}
}
