% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category.products.and.services.R
\name{category.products.and.services}
\alias{category.products.and.services}
\title{Category Products and Services Search Parameter}
\usage{
category.products.and.services(...)
}
\arguments{
\item{...}{One or more keywords.}
}
\value{
An S3 search parameter object.
}
\description{
Category products and services search parameters for a targeting idea.
See \url{https://developers.google.com/adwords/api/docs/reference/v201809/TargetingIdeaService.CategoryProductsAndServicesSearchParameter}.
}
\examples{
# Search parameter for vertical Advertising & Marketing
category.products.and.services(25)
}
