% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.report.R
\name{check.report}
\alias{check.report}
\title{Adwords Report Error Handling}
\usage{
check.report(x)
}
\arguments{
\item{x}{The result from a report request.}
}
\value{
An error message if an error is found, or NULL if none were found.
}
\description{
Checks a report response in the \link{get.report} function.
If the response is a data.frame then the request worked.
If the response is still in XML than an error occured.
The XML is parsed to extract and return the error message.
}
\examples{
# Try to get a non existing report.
library(RAdwords)
report <- "INVALID_REPORT"
cid <- "XXX-XXX-XXXX"
auth <- doAuth()
fields <- c("AccountDescriptiveName", "AdGroupId", "AdGroupName", "AdGroupStatus", "CampaignId", "CampaignName", "CampaignStatus", "KeywordId", "KeywordTextMatchingQuery", "QueryMatchTypeWithVariant", "Impressions", "Clicks", "Conversions", "Cost", "AveragePosition")
report <- get.report(report, cid, auth, fields, date = "LAST_14_DAYS")
}
\seealso{
get.report
}
