% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.service.R
\name{check.service}
\alias{check.service}
\title{Adwords Service Error Handling}
\usage{
check.service(x)
}
\arguments{
\item{x}{The result from a service request.}
}
\value{
An error message if an error is found, or NULL if none were found.
}
\description{
Checks a service response in the \link{get.service} function.
The XML is parsed to detect an error an extract the error message.
Then the error message is returned.
}
\examples{
TODO
}
\seealso{
get.service
}

