% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.traffic.estimator.R
\name{combine.traffic.estimator}
\alias{combine.traffic.estimator}
\title{Combine Adwords Traffic Request and Estimate}
\usage{
combine.traffic.estimator(cer, estimate)
}
\arguments{
\item{estimate}{An object of type traffic.estimator.}

\item{request}{An object of type campaign.estmate.request}
}
\value{
A data frame with the combined informations.
}
\description{
Combines a \link{campaign.estmate.request} and the \link{traffic.estimator} obtained with this campaign.estimate.request into one final data.frame.
In the Adwords Traffict Estimate, the requested keywords are not returned.
This method parse the returned estimates and join the keywords from the request.
The solution uses the fact that the results in the estimate are in the same order as the keywords in the request.
}
\examples{
# For this exemple to work, you must supply a valid client customer id and your developper token
library(RAdwords)
k <- keyword(c("Example", "Test"), match.type = c("EXACT", "PHRASE"))
ker <- keyword.estimate.request(k, max.cpc = c(1500000, 2250000))
aer <- adgroup.estimate.request(ker)
cer <- campaign.estimate.request(aer, campaign.id = "123456789")
request <- traffic.estimator.request(cer)
estimate <- get.service(request, cid = "YOUR_CID", auth = doAuth(), cid = "YOUR_CID", dev.token = "YOUR_DEV_TOKEN", user.agent = "AKTE Exemple")
combine.traffic.estimator(cer, estimate)
}

