% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{criterion}
\alias{as.criterion}
\alias{criterion}
\alias{is.criterion}
\title{Adwords Criterion}
\usage{
criterion(id, type = "Location")

as.criterion(id, ...)

is.criterion(x)
}
\arguments{
\item{id}{Amount to bid for the maximum cost per click.}

\item{type}{TRUE to specify that the amount x must be converted to a microamount, FALSE to avoid any conversion for bids that are already micromounts. Default to FALSE.}
}
\value{
A S3 object of type criterion.
}
\description{
Creates or coerces S3 objects of type \code{criterion} for the Adwords API.
\code{is.criterion} will test if an object is interpretable as a criterion.
Criteria can be added to a \code{campaign.estimate.request}.
Only locations and languages are supported in the keyword traffic estimator.
}
\examples{
usa <- criterion(2840) # United States
english <- criterion("1000", type = "Language") # English
french <- as.criterion(1002, type = "Language") # French
is.criterion(french)
}

