% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.sync.parser.R
\name{customer.sync.parser}
\alias{changed.adgroups.parser}
\alias{changed.campaigns.parser}
\alias{customer.sync.parser}
\title{XML Parsing for Adwords Customer Sync}
\usage{
customer.sync.parser(x, ...)

changed.campaigns.parser(node)

changed.adgroups.parser(node)
}
\arguments{
\item{x}{Customer sync response in XML.}
}
\value{
A list of two data.frames with the entries and the links.
}
\description{
Parses an Adwords customer sync service response in XMl format into a TODO.
}
\examples{
# For this exemple to work, you must supply a valid client customer id and your developper token
TODO
}

