% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.sync.request.R
\name{customer.sync.request}
\alias{customer.sync.request}
\title{Adwords Customer Sync Request}
\usage{
customer.sync.request(campaign.ids, feed.ids = NULL, min.date.time = NULL,
  max.date.time = NULL)
}
\arguments{
\item{campaign.ids}{Vector or list of campaign ids for which we want the changes.}

\item{min.date.time}{The min date time of the range for the changes in yyyyMMdd HHmmss <Timezone ID> format. Default to system time if NULL.}

\item{max.date.time}{The min date time of the range for the changes in yyyyMMdd HHmmss <Timezone ID> format. Default to system time if NULL.}

\item{campaign.ids}{Vector or list of feed ids for which we want the changes.}
}
\value{
An XML formatted customer sync request.
}
\description{
Prepares the body of an XML formatted request for the Adwords customer sync service.
}
\examples{
request <- customer.sync.request(campaign.ids = c("123456789", "456789123", "789123456"))
}
\seealso{
get.service
}

