% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissect.R
\name{dissect}
\alias{dissect}
\alias{dissect.entry}
\alias{dissect.link}
\alias{dissect.list}
\alias{dissect.managed.customer}
\title{XML Parsing for Adwords}
\usage{
dissect(x, ...)

\method{dissect}{list}(x, ...)

\method{dissect}{managed.customer}(x, ...)

\method{dissect}{entry}(entry.node)

\method{dissect}{link}(entry.node)
}
\arguments{
\item{x}{Managed Customer response in XML.}
}
\value{
A data.frame with the estimate results.
}
\description{
Dissects an Adwords service response in XMl format into the appropriate format for the results.
}
\examples{
# For this exemple to work, you must supply a valid client customer id and your developper token
TODO
}

