% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.service.R
\name{get.service}
\alias{get.service}
\title{Call an Adwords Service}
\usage{
get.service(request, cid, auth, user.agent, api.version = "v201702",
  validate.only = FALSE, partial.failure = FALSE, verbose = FALSE,
  raw = FALSE)
}
\arguments{
\item{request}{XMLNode for the body of an XML encoded Adwords service request.}

\item{cid}{client customer id.}

\item{auth}{Google authentification made with \code{\link{doAuth}}.}

\item{user.agent}{User agent string for your script, read \url{http://googleadsdeveloper.blogspot.ca/2013/11/please-set-user-agent-or-application.html} for more information.}

\item{api.version}{version of the Adwords API to use, must be the same as in the request. Default to version v201702.}

\item{validate.only}{Test the API call without actually executing the call against real data. Default to FALSE.}

\item{partial.failure}{If true service will carry out the operations that had no errors. Default to FALSE.}

\item{verbose}{TRUE to turn on the verbose mode of the \code{\link{getURL}} call. Default to FALSE.}

\item{raw}{TRUE to return the raw XML response. Default to FALSE.}
}
\value{
The XML response for the service request.
}
\description{
Sends a request for a Google Adwords service.
}
\examples{
# For this exemple to work, you must supply a valid client customer id and your developper token
data <- get.service(request, cid = cid, auth = doAuth(), user.agent = user.agent)
}
\seealso{
\code{\link{managed.customer.request}}
}
