% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.error.R
\name{is.error}
\alias{is.error}
\alias{is.error.data.frame}
\alias{is.error.default}
\alias{is.error.list}
\title{Check Adwords API Response for Error}
\usage{
is.error(x)

\method{is.error}{list}(x, ...)

\method{is.error}{data.frame}(x)

\method{is.error}{default}(x)
}
\arguments{
\item{x}{Adwords XML response.}
}
\value{
Logical. TRUE if this XML message is an error.
}
\description{
Checks if the Adwords API response is an error.
}
\examples{
xml.error.msg <- "<?xml version=\\"1.0\\" encoding=\\"UTF-8\\" standalone=\\"yes\\"?><reportDownloadError><ApiError><type>ReportDefinitionError.CUSTOMER_SERVING_TYPE_REPORT_MISMATCH</type><trigger></trigger><fieldPath>selector.reportDefinition</fieldPath></ApiError></reportDownloadError>"
is.error(xml.error.msg)
}

