% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword.estimate.request.R
\name{keyword.estimate.request}
\alias{as.keyword.estimate.request}
\alias{as.keyword.estimate.request.data.frame}
\alias{as.keyword.estimate.request.default}
\alias{is.keyword.estimate.request}
\alias{keyword.estimate.request}
\title{Adwords Keyword Estimate Request}
\usage{
keyword.estimate.request(keyword, max.cpc = NULL, is.negative = NULL)

as.keyword.estimate.request(x, ...)

\method{as.keyword}{estimate.request.data.frame}(x, max.cpc = NULL,
  is.negative = NULL, ...)

\method{as.keyword}{estimate.request.default}(x, ...)

is.keyword.estimate.request(x)
}
\arguments{
\item{keyword}{The keyword objects to prepare into a keyword.estimate.request.}

\item{max.cpc}{The bid for the keyword.}

\item{is.negative}{When TRUE, no current ad group ads will appear for searches containing this keyword.}
}
\value{
A S3 of type keyword.estimate.request.
}
\description{
Creates or coerces S3 objects of type \code{keyword.estimate.request} for the Adwords API.
\code{is.keyword.estimate.request} will test if an object is interpretable as a keyword.estimate.request
}
\examples{
# Minimal example
k1 <- keyword("Example")
request1 <- keyword.estimate.request(k1)

# More complex example
k2 <- keyword(c("Example", "Sample", "Test", "Only a test"), match.type = "PHRASE")
request2 <- keyword.estimate.request(k2, max.cpc = max.cpc(c(1, 2, 1, 3), use.micro = TRUE))

# Coercing to a keyword.estimate.request
k3 <- as.keyword.estimate.request("Convert", is.negative = TRUE, match.type = "EXACT")

# Checking a keyword.estimate.request
is.keyword.estimate.request(k3)
}

