% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword.R
\name{keyword}
\alias{keyword}
\alias{as.keyword}
\alias{as.keyword.data.frame}
\alias{as.keyword.default}
\alias{is.keyword}
\title{Adwords Keyword}
\usage{
keyword(text, match.type = "BROAD", keyword.id = NULL,
  type = "Keyword")

as.keyword(x, ...)

\method{as.keyword}{data.frame}(x, keyword.id, match.type = NULL,
  type = "Keyword")

\method{as.keyword}{default}(x, ...)

is.keyword(x)
}
\arguments{
\item{text}{Text of the keyword.}

\item{match.type}{Either "BROAD", "PHRASE" or "EXACT".}

\item{keyword.id}{The ID for the keyword, unsuported in the Adwords keyword traffic estimator service. When converting a data frame, this parameter will override the existing column with keyword ids, remove it if NULL or leave the column intact if missing (which is the default).}

\item{type}{Placeholder for the keyword type, unsuported for now.}
}
\value{
A S3 object of type keyword.
}
\description{
Creates or coerces S3 objects of type \code{keyword} for the Adwords API.
\code{is.keyword} will test if an object is interpretable as a keyword.
}
\examples{
# Simplest case
k1 <- keyword("Example")

# Adding a match type
k2 <- keyword("Example", "EXACT")

# Creating multiple keywords
k3 <- keyword(c("Example", "Sample", "Test", "Only a test"), match.type = "PHRASE")

# Creating multiple keywords with different match types
k4 <- keyword(c("Example", "Sample", "Test", "Only a test"), match.type = c("PHRASE", "BROAD", "EXACT", "BROAD"))

# Coercing to a keyword
k5 <- as.keyword(911)

# Checking a keyword
is.keyword(k5)
}
