% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{language}
\alias{language}
\title{Language Search Parameter}
\usage{
language(...)
}
\arguments{
\item{...}{One or more language id.}
}
\value{
An S3 search parameter object.
}
\description{
Language search parameters for a targeting idea.
See \url{https://developers.google.com/adwords/api/docs/reference/v201702/TargetingIdeaService.LanguageSearchParameter}.
}
\examples{
# English and French
language(1000, 1002)
}
