% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\name{location}
\alias{location}
\title{Location Search Parameter}
\usage{
location(...)
}
\arguments{
\item{...}{One or more location id.}
}
\value{
An S3 search parameter object.
}
\description{
Location search parameters for a targeting idea.
See \url{https://developers.google.com/adwords/api/docs/reference/v201702/TargetingIdeaService.LocationSearchParameter}.
}
\examples{
# USA and Canada
location(2840, 2124)
}

