% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managed.customer.parser.R
\name{managed.customer.parser}
\alias{managed.customer.entry.parser}
\alias{managed.customer.link.parser}
\alias{managed.customer.parser}
\title{XML Parsing for Adwords Managed Customer}
\usage{
managed.customer.parser(x, ...)

managed.customer.entry.parser(entry.node)

managed.customer.link.parser(entry.node)
}
\arguments{
\item{x}{Managed customer response in XML.}
}
\value{
A list of two data.frames with the entries and the links.
}
\description{
Parses an Adwords managed customer service response in XMl format into a list of two data.frames : the customer account entries and the management links.
}
\examples{
# For this exemple to work, you must supply a valid client customer id and your developper token
TODO
}

