% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managed.customer.request.R
\name{managed.customer.request}
\alias{managed.customer.request}
\title{Adwords Managed Customer Service Request}
\usage{
managed.customer.request(fields = c("Name", "CustomerId"),
  predicates = NULL, start.index = 0, number.of.results = 100)
}
\arguments{
\item{fields}{List of fields to select.}

\item{predicates}{Specifies how an entity (eg. adgroup, campaign, criterion, ad) should be filtered.}

\item{start.index}{Index of the first result to return with this request. Default is 0.}

\item{number.of.results}{Maximum number of results to return in this request. Default is 100.}
}
\value{
An XML formatted managed customer request.
}
\description{
Prepares the body of an XML formatted request for the Adwords managed customer service.
Only data retrieval is implemented for now.
}
\examples{
# Prepare a request to get all account in a MCC that starts with "USA" and doesn't contain "inc".
preds <- list(predicate("Name", "STARTS_WITH_IGNORE_CASE", "USA"), predicate("Name", "DOES_NOT_CONTAIN_IGNORE_CASE", "inc"))
request <- managed.customer.request(fields = c("Name", "CustomerId"), predicates = preds)
}
\seealso{
get.service
}
