% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.cpc.R
\name{max.cpc}
\alias{as.max.cpc}
\alias{is.max.cpc}
\alias{max.cpc}
\title{Adwords Maximum Cost per Click}
\usage{
\method{max}{cpc}(bid, use.micro = FALSE)

as.max.cpc(bid, ...)

is.max.cpc(x)
}
\arguments{
\item{bid}{Amount to bid for the maximum cost per click.}

\item{use.micro}{TRUE to specify that the amount x must be converted to a microamount, FALSE to avoid any conversion for bids that are already micromounts. Default to FALSE.}
}
\value{
A S3 object of type max.cpc.
}
\description{
Creates or coerces S3 objects of type \code{max.cpc} for the Adwords API.
\code{is.max.cpc} will test if an object is interpretable as a max.cpc.
}
\examples{
bid1 <- max.cpc(0.99)
bid2 <- max.cpc(990000, use.micro = TRUE)
bid3 <- as.max.cpc("0.99")
is.max.cpc(bid3)
}

