% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.error.R
\name{parse.error}
\alias{parse.error}
\title{XML Error Parsing for Adwords}
\usage{
parse.error(x, ...)
}
\arguments{
\item{x}{Adwords API response in XML.}
}
\value{
A data.frame with the estimate results.
}
\description{
Parses an error message in XMl format into a message explaining the error.
}
\examples{
# For this exemple to work, you must supply a valid client customer id and your developper token
library(RAdwords)
TODO
}

