% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.managed.customer.R
\name{parse.managed.customer}
\alias{parse.managed.customer}
\alias{parse.managed.customer.entry}
\alias{parse.managed.customer.link}
\title{XML Parsing for Adwords Managed Customer}
\usage{
parse.managed.customer(x, ...)

parse.managed.customer.entry(entry.node)

parse.managed.customer.link(entry.node)
}
\arguments{
\item{x}{Managed customer response in XML.}
}
\value{
A list of two data.frames with the entries and the links.
}
\description{
Parses an Adwords managed customer service response in XMl format into a list of two data.frames : the customer account entries and the management links.
}
\examples{
# For this exemple to work, you must supply a valid client customer id and your developper token
TODO
}

