% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.targeting.idea.R
\name{parse.targeting.idea}
\alias{parse.targeting.idea}
\alias{parse.targeting.idea.data}
\alias{parse.targeting.idea.entry}
\title{XML Parsing for Adwords Targeting Idea}
\usage{
parse.targeting.idea(x, ...)

parse.targeting.idea.entry(entry.node)

parse.targeting.idea.data(data.node)
}
\arguments{
\item{x}{Targeting idea response in XML.}
}
\value{
A list of two data.frames with the entries and the links.
}
\description{
Parses an Adwords targeting idea service response in XMl format into a data.frame.
}
\examples{
# For this exemple to work, you must supply a valid client customer id and your developper token
TODO
}

