% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.traffic.estimator.R
\name{parse.traffic.estimator}
\alias{parse.adgroup.estimate}
\alias{parse.campaign.estimate}
\alias{parse.keyword.estimate}
\alias{parse.traffic.estimator}
\title{XML Parsing for the Traffic Estimator Service.}
\usage{
parse.traffic.estimator(x, ...)

parse.campaign.estimate(node, default.id = NULL)

parse.adgroup.estimate(node, default.id = NULL)

parse.keyword.estimate(node, default.id = NULL)
}
\arguments{
\item{x}{traffic.estimator response in XML.}
}
\value{
A data.frame with the estimate results.
}
\description{
Parses a \link{traffic.estimator} response in XMl format into a data frame containing the results of the request.
The data frame is in long format.
Use dcast in package reshape2 with \code{formula = CampaignID + AdgroupID + KeywordID ~ Metric} and \code{value.var = "Value"} to get a data.frame in wide format.
}
\details{
There's also a parse function for \link{traffic.estimator.request} to convert the requests a list of \link{campaign.estimate.request}.
}
\examples{
# For this exemple to work, you must supply a valid client customer id and your developper token
library(RAdwords)
k <- keyword(c("Example", "Test"), match.type = c("EXACT", "PHRASE"))
ker <- keyword.estimate.request(k, max.cpc = c(1500000, 2250000))
aer <- adgroup.estimate.request(ker)
cer <- campaign.estimate.request(aer, campaign.id = "123456789")
request <- traffic.estimator.request(cer)
estimate <- get.service(request, cid = "YOUR_CID", auth = doAuth(), dev.token = "YOUR_DEV_TOKEN", user.agent = "AKTE Exemple")

# Parse the request to get back the data
parse(request)

# Parse the estimate to get the results
parse.traffic.estimator(estimate)
}

