% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/related.to.query.R
\name{related.to.query}
\alias{related.to.query}
\title{Related to Query Search Parameter}
\usage{
related.to.query(...)
}
\arguments{
\item{...}{One or more keywords.}
}
\value{
An S3 search parameter object.
}
\description{
Related to query search parameters for a targeting idea.
See \url{https://developers.google.com/adwords/api/docs/reference/v201710/TargetingIdeaService.RelatedToQuerySearchParameter}.
}
\examples{
# Queries for some planets
related.to.query("Mercury", "Venus", "Mars")
}
