% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Retrieve an Adwords Report}
\usage{
report(cid, auth, name = "ACCOUNT_PERFORMANCE_REPORT",
  fields = c("AccountDescriptiveName", "Impressions", "Clicks", "Cost",
  "Date"), date = "LAST_14_DAYS", ...)
}
\arguments{
\item{cid}{Customer client ids (or vector of).}

\item{auth}{Google AdWords authentification token.}

\item{name}{Name of the report. Default to the account performance report ("ACCOUNT_PERFORMANCE_REPORT").}

\item{fields}{Metrics to select in the report. This can be a mix of attributes, segments and metrics. Default to "AccountDescriptiveName", "Impressions", "Clicks", "Cost", "Date".}

\item{date}{Either a date range (see \url{https://developers.google.com/adwords/api/docs/guides/reporting#date-ranges}) or a vector of length two with the date interval (start date first, end date last). Default to last fourteen days ("LAST_14_DAYS")}

\item{...}{Extra parameters for the RAdwords::statement function}
}
\value{
The report as a data.table with the selected columns.
}
\description{
Downloads an Adwords report. Wrapper function around the \link[RAdwords]{statement} and \link[RAdwords]{getData} functions.
}
\examples{
# Download the search query performance report for January 2015
auth <- doAuth()
name <- "SEARCH_QUERY_PERFORMANCE_REPORT"
fields <- c("AccountDescriptiveName", "AdGroupId", "AdGroupName", "AdGroupStatus", "CampaignId", "CampaignName", "CampaignStatus", "KeywordId", "KeywordTextMatchingQuery", "QueryMatchTypeWithVariant", "Impressions", "Clicks", "Conversions", "Cost", "AveragePosition")
cid <- c('XXX-XXX-XXXX', 'YYY-YYY-YYYY')
report <- adwords.report(cid, auth, name, fields, date = "LAST_14_DAYS")
}

