% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeting.idea.parser.R
\name{targeting.idea.parser}
\alias{targeting.idea.data.parser}
\alias{targeting.idea.entry.parser}
\alias{targeting.idea.parser}
\title{XML Parsing for Adwords Targeting Idea}
\usage{
targeting.idea.parser(x, ...)

targeting.idea.entry.parser(entry.node)

targeting.idea.data.parser(data.node)
}
\arguments{
\item{x}{Targeting idea response in XML.}
}
\value{
A list of two data.frames with the entries and the links.
}
\description{
Parses an Adwords targeting idea service response in XMl format into a data.frame.
}
\examples{
# For this exemple to work, you must supply a valid client customer id and your developper token
TODO
}

