% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.estimator.request.R
\name{traffic.estimator.request}
\alias{is.traffic.estimator.request}
\alias{traffic.estimator.request}
\title{Adwords Traffic Estimator Request}
\usage{
traffic.estimator.request(cer, cid, dev.token, user.agent,
  api.version = "v201702", validate.only = FALSE, partial.failure = FALSE)

is.traffic.estimator.request(x)
}
\arguments{
\item{cer}{campaign.estimate.request to encode.}
}
\value{
An XML formatted estimate request.
}
\description{
Encodes a \code{\link{campaign.estimate.request}} into a an XML formatted request for the Adwords Keyword Traffic Estimator.
\code{is.traffic.estimator.request} will test if an object is interpretable as a traffic.estimator.request.
}
\examples{
# For this exemple to work, you must supply a valid client customer id and your developper token
k <- keyword(c("Example", "Test"), match.type = c("EXACT", "PHRASE"))
ker <- keyword.estimate.request(k, max.cpc = c(1500000, 2250000))
aer <- adgroup.estimate.request(ker)
cer <- campaign.estimate.request(aer, campaign.id = "123456789")
request <- traffic.estimator.request(cer, cid = "YOUR_CID", dev.token = "YOUR_DEV_TOKEN", user.agent = "AKTE Exemple")
}

