% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adwords.fields}
\alias{adwords.fields}
\title{Adwords Report Fields}
\format{A data frame with 7830 rows and 11 variables}
\usage{
adwords.fields
}
\description{
A table containing all the Adwords report fields (attributes, segments and metrics) for the 201609, 201607 and 201605 API versions.
}
\keyword{Adwords}
\keyword{Attribute}
\keyword{Field}
\keyword{Metric}
\keyword{Segment}
\keyword{datasets}

