% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.xml.R
\name{as.xml}
\alias{as.xml}
\alias{as.xml.NULL}
\alias{as.xml.list}
\alias{as.xml.logical}
\alias{as.xml.predicate}
\title{XML Encoding for Adwords Service Requests}
\usage{
as.xml(x)

\method{as.xml}{NULL}(x)

\method{as.xml}{list}(x)

\method{as.xml}{logical}(x)

\method{as.xml}{predicate}(x)
}
\arguments{
\item{x}{Data to convert to XML.}
}
\value{
An XMLNode encoding the data
}
\description{
Encodes the search parameters of an Adwords service request into XMl.
Allow the conversion of the data to a SOAP message for the Adwords service.
}
\examples{
TODO
}

