% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statement.R
\name{statement}
\alias{statement}
\title{Prepare an Adwards Query Language Statement}
\usage{
statement(report = "ACCOUNT_PERFORMANCE_REPORT",
  fields = c("AccountDescriptiveName", "Impressions", "Clicks", "Cost",
  "Date"), date = "LAST_14_DAYS", where)
}
\arguments{
\item{report}{Name of the report. Default to the account performance report ("ACCOUNT_PERFORMANCE_REPORT").}

\item{fields}{Fields to select in the report. This can be a mix of attributes, segments and metrics. Default to "AccountDescriptiveName", "Impressions", "Clicks", "Cost", "Date".}

\item{date}{Either a date range (see \url{https://developers.google.com/adwords/api/docs/guides/reporting#date-ranges}) or a vector of length two with the date interval (start date first, end date last). Default to last fourteen days ("LAST_14_DAYS"). Note that this works only if no Date or Week columns are specified (see \url{https://developers.google.com/adwords/api/docs/guides/awql}).}

\item{where}{Conditions to use in the statement, like restrictions on a field}
}
\value{
A character string with the content of the query.
}
\description{
Builds an Adwards Query Language Statement.
}
\examples{
statement(report = "ACCOUNT_PERFORMANCE_REPORT", fields = c("AccountDescriptiveName", "Impressions", "Clicks", "Cost", "Date"), date = "LAST_14_DAYS")
}

